[FEAT] : 맵뷰 필터바텀시트뷰 카테고리버튼 수정 리스트뷰 레이아웃 업데이트

+++ PopPool/PopPool/Data/Network/PopPoolNetwork/MapDTO/PopUpStoreDTO.swift
@@ -7,13 +7,14 @@ struct PopUpStoreDTO: Codable {
     let category: String
     let name: String
     let address: String
-    let startDate: String  // ISO 8601 형식의 날짜 문자열
-    let endDate: String    // ISO 8601 형식의 날짜 문자열
+    let startDate: String
+    let endDate: String    
     let latitude: Double
     let longitude: Double
-    let markerId: Int
+    let markerId: Int64  
     let markerTitle: String
     let markerSnippet: String
+    
 
     func toDomain() -> PopUpStore {
         return PopUpStore(
